+++ src/find_next_move.c
@@ -24,7 +24,7 @@ MCTSNode* expandLeaf(Board* board, MCTSNode* leaf) {
 Winner simulate(Board* board, pcg32_random_t* rng) {
     while (getWinner(board) == NONE) {
         Square validMoves[TOTAL_SMALL_SQUARES];
-        int amountOfMoves = generateMoves(board, validMoves);
+        uint8_t amountOfMoves = generateMoves(board, validMoves);
         Square move = validMoves[pcg32_boundedrand_r(rng, amountOfMoves)];
         makeTemporaryMove(board, move);
     }
